+++ tests/unit/views/account_test.py
@@ -8,7 +8,7 @@ import pytest
 from funnel import models
 from funnel.views.account import user_agent_details
 
-from ...conftest import LoginFixtureProtocol, TestClient, scoped_session
+from ...conftest import TestClient
 
 
 def test_account_always_has_profile_url(
@@ -22,7 +22,6 @@ def test_account_always_has_profile_url(
 
 
 def test_username_available(
-    db_session: scoped_session,
     client: TestClient,
     user_rincewind: models.User,
     csrf_token: str,
@@ -71,60 +70,6 @@ def test_username_available(
         'error_description': "Usernames can only have alphabets, numbers and"
         " underscores",
     }
-
-
-# Sample password that will pass zxcvbn's complexity validation, but will be flagged
-# by the pwned password validator
-PWNED_PASSWORD = 'thisisone1'  # nosec
-
-
-@pytest.mark.enable_socket()
-def test_pwned_password(
-    client: TestClient,
-    csrf_token: str,
-    login: LoginFixtureProtocol,
-    user_rincewind: models.User,
-) -> None:
-    """Pwned password validator will block attempt to use a compromised password."""
-    login.as_(user_rincewind)
-    rv = client.post(
-        'account/password',
-        data={
-            'username': user_rincewind.username,
-            'form.id': 'password-change',
-            'password': PWNED_PASSWORD,
-            'confirm_password': PWNED_PASSWORD,
-            'csrf_token': csrf_token,
-        },
-    )
-    assert rv.status_code == 200
-    assert "This password was found in breached password lists" in rv.data.decode()
-
-
-def test_pwned_password_mock_endpoint_down(
-    requests_mock,
-    client: TestClient,
-    csrf_token: str,
-    login: LoginFixtureProtocol,
-    user_rincewind: models.User,
-) -> None:
-    """If the pwned password API is not available, the password is allowed."""
-    requests_mock.get('https://api.pwnedpasswords.com/range/1F074', status_code=404)
-    login.as_(user_rincewind)
-
-    rv = client.post(
-        'account/password',
-        data={
-            'username': user_rincewind.username,
-            'form.id': 'password-change',
-            'password': PWNED_PASSWORD,
-            'confirm_password': PWNED_PASSWORD,
-            'csrf_token': csrf_token,
-        },
-    )
-
-    assert rv.status_code == 303
-    assert rv.location == '/account'
 
 
 @pytest.mark.parametrize(
